build - implemented transaction notifications via email and fixed multiple issues with withdrawals and deposits

+++ transaction-service/src/main/java/com/example/transactionservice/client/AccountClient.java
@@ -9,5 +9,5 @@ import org.springframework.web.bind.annotation.PathVariable;
 public interface AccountClient {
 
     @GetMapping("/{id}")
-    AccountDetails findAccount(@PathVariable int id);
+    AccountDetails findAccount(@PathVariable Long id);
 }
